+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexUnknownConfigPropertyInspection.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -28,7 +28,7 @@ import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroDeclaration
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroValue
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexVisitor
-import com.intellij.idea.plugin.hybris.impex.utils.ProjectPropertiesUtils
+import com.intellij.idea.plugin.hybris.properties.PropertiesService
 import com.intellij.lang.properties.PropertiesImplUtil
 import com.intellij.psi.PsiElementVisitor
 import com.intellij.psi.util.PsiTreeUtil
@@ -56,7 +56,9 @@ private class UnknownConfigPropertyVisitor(private val problemsHolder: ProblemsH
                     ProblemHighlightType.ERROR
                 )
             } else {
-                val property = ProjectPropertiesUtils.findMacroProperty(usage.project, propertyName)
+                val propertiesService = PropertiesService.getInstance(usage.project) ?: return
+                val property = propertiesService.findMacroProperty(usage.project, propertyName)
+
                 if (property == null) {
                     cachedProperties[propertyName] = false
                     problemsHolder.registerProblem(