+++ src/com/intellij/idea/plugin/hybris/impex/completion/provider/ImpexMacrosConfigCompletionProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,7 +23,7 @@ import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.completion.CompletionUtilCore
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
-import com.intellij.idea.plugin.hybris.impex.utils.ProjectPropertiesUtils
+import com.intellij.idea.plugin.hybris.properties.PropertiesService
 import com.intellij.idea.plugin.hybris.system.type.codeInsight.lookup.TSLookupElementFactory
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.psi.PsiElement
@@ -34,11 +34,14 @@ class ImpexMacrosConfigCompletionProvider : CompletionProvider<CompletionParamet
 
     override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
         val psiElementUnderCaret = parameters.position
+        val project = psiElementUnderCaret.project
         val prevLeaf = PsiTreeUtil.prevLeaf(psiElementUnderCaret)
+        val propertiesService = PropertiesService.getInstance(project) ?: return
+
         if (prevLeaf != null && prevLeaf.text.contains(HybrisConstants.IMPEX_CONFIG_COMPLETE_PREFIX)) {
             val position = parameters.position
             val query = getQuery(position)
-            ProjectPropertiesUtils.findAutoCompleteProperties(position.project, query)
+            propertiesService.findAutoCompleteProperties(query)
                 .mapNotNull { it.key }
                 .map { TSLookupElementFactory.buildCustomProperty(it) }
                 .forEach { result.addElement(it) }
@@ -50,7 +53,7 @@ class ImpexMacrosConfigCompletionProvider : CompletionProvider<CompletionParamet
             val query = position.text
                 .substring(prefix.length)
                 .replace(CompletionUtilCore.DUMMY_IDENTIFIER_TRIMMED, "")
-            ProjectPropertiesUtils.findAutoCompleteProperties(position.project, query)
+            propertiesService.findAutoCompleteProperties(query)
                 .mapNotNull { it.key }
                 .map { it }
                 .map { TSLookupElementFactory.buildCustomProperty(it) }