+++ src/com/intellij/idea/plugin/hybris/impex/psi/references/ImpexPropertyReference.kt
@@ -1,10 +1,11 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -20,7 +21,7 @@ package com.intellij.idea.plugin.hybris.impex.psi.references
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
-import com.intellij.idea.plugin.hybris.impex.utils.ProjectPropertiesUtils
+import com.intellij.idea.plugin.hybris.properties.PropertiesService
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElementResolveResult
 import com.intellij.psi.PsiReference
@@ -44,10 +45,15 @@ class ImpexPropertyReference(owner: ImpexMacroUsageDec) : PsiReferenceBase.Poly<
 
     override fun getVariants(): Array<PsiReference> = PsiReference.EMPTY_ARRAY
 
-    override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> = getPropertyKey()
-        ?.let { ProjectPropertiesUtils.findMacroProperty(element.project, it) }
-        ?.let { PsiElementResolveResult.createResults(it.psiElement) }
-        ?: ResolveResult.EMPTY_ARRAY
+    override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {
+        val propertiesService = PropertiesService.getInstance(element.project)
+            ?: return emptyArray()
+
+        return getPropertyKey()
+            ?.let { propertiesService.findMacroProperty(element.project, it) }
+            ?.let { PsiElementResolveResult.createResults(it.psiElement) }
+            ?: ResolveResult.EMPTY_ARRAY
+    }
 
     private fun getPropertyKey() = element.text
         .replace(HybrisConstants.IMPEX_CONFIG_COMPLETE_PREFIX, "")