+++ src/com/intellij/idea/plugin/hybris/system/type/codeInsight/completion/TSCompletionService.kt
@@ -39,7 +39,7 @@ interface TSCompletionService {
     /**
      * See: https://help.sap.com/docs/SAP_COMMERCE/d0224eca81e249cb821f2cdf45a82ace/2fb5a2a780c94325b4a48ff62b36ab23.html#using-header-abbreviations
      */
-    fun getHeaderAbbreviationCompletions(): List<LookupElement>
+    fun getHeaderAbbreviationCompletions(project: Project): List<LookupElement>
 
     companion object {
         fun getInstance(project: Project): TSCompletionService = project.getService(TSCompletionService::class.java)